Make G-vector routines orders of magnitude faster
We now use an auxiliary list to quickly find rotated G-vectors.
This makes all G-vectors operations scale as `O(N_G)` instead of `O(N_G^2)`,
where `N_G` is the number of G-vectors.

We also add a flag in `loadumat.f90` to skip reading the Umat file and using
the identity matrix insted. This should really be an input flag!